refactor: 놀이터 참여 response 데이터 변경
